% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batchSfile.r
\name{batch.sfile}
\alias{batch.sfile}
\title{Batch Read Winsteps Structure Files (sfiles) Into R}
\usage{
batch.sfile(dir = getwd(), pat = "Sfile", files = NULL)
}
\arguments{
\item{dir}{The directory where the item files are saved.}

\item{pat}{A common character string in all item files. Only used if
\code{files = NULL}, Defaults to "Ifile", which is what is automatically
generated when \code{link{r2Winsteps}} is called.}

\item{files}{Optional character vector of the names of the item files to be
  read into R. Note that only the names of the files (not the extension) are
  neccessary. The function assumes .txt files.}
}
\value{
List of Winsteps sfiles, where each element of the list represents
a different sfile.
}
\description{
This function is essentially equivalent to \code{\link{batch.ifile}},
except that structure files are read into R rather than item files. These
are necessary for plotting results from rating scale or partial credit
models.
}

