% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_r2Winsteps.r
\name{batch_r2Winsteps}
\alias{batch_r2Winsteps}
\title{Write Winsteps control and data files as a batch.}
\usage{
batch_r2Winsteps(itmsL, demsL, batName = "r2WinstepsBatch",
  partialCreditL = NULL, anchorFileL = NULL, titleL = NULL, ...)
}
\arguments{
\item{itmsL}{List of dataframes or matrices containing only the item 
responses. Each element of the list should contain a seperate dataframe to
be processed.}

\item{demsL}{List of dataframes or matrices containing person 
identifiers/demographic fields.}

\item{...}{Additional argument supplied to \code{\link{r2Winsteps}}. Note 
that any additional arguments must be supplied in the form of a list.}
}
\value{
Control and data files for all dataframes/matrices in the given 
  lists, as well as a .bat file to batch process the files in Winsteps.
}
\description{
Function takes the typical arguments supplied by \code{\link{r2Winsteps}}
 in the form of a list, and writes out Winsteps control and data files for 
 each element of the list. See also \code{\link{r2Winsteps}}.
}

