% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batchRunWinsteps.r
\name{batchRunWinsteps}
\alias{batchRunWinsteps}
\title{Run a Rasch model from R using Winsteps.}
\usage{
batchRunWinsteps(itmsL, demsL, titleL = NULL, keep = FALSE, ...)
}
\arguments{
\item{itmsL}{List of dataframes or matrices of item responses to be analyzed.}

\item{demsL}{List of dataframes or matrices of person identifiers/demographic
fields.}

\item{titleL}{Optional list of names for control and datafiles written to the
working directory. Note that this argument is only neccessary if
\code{keep = TRUE}.}

\item{keep}{Should the external files used to conduct the analysis be stored
in the working directory? If TRUE, all Winsteps files will be retained,
including control, data, item, and person files in .txt format, as well as
the .bat file used to process the data. Defaults to FALSE, in which case
no files are retained. Alternatively, a character vector of the files to
retain can be supplied. Possible arguments are \code{c("iFile", "pFile",
"cntrlFile", "dtaFile", "outFile", "bat")}. These correspond to,
respectively, item, person, control, data, output, and the .bat (batch
processing) files.}

\item{...}{Additional arguments passed to \code{\link{batch_r2Winsteps}}.}
}
\value{
List containing all item and person parameters from the given
  analyses.
}
\description{
This function uses the Winsteps software to batch analyze a given list of
  dataframes/matrices. The results are read back into R without leaving the
  R interface or interacting with the Winsteps GUI.
}
\seealso{
\code{\link{batch_r2Winsteps}} \code{\link{batchWinsteps}}
  \code{\link{batch.pfile}} \code{\link{batch.ifile}}
}

