% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batchWinsteps.r
\name{batchWinsteps}
\alias{batchWinsteps}
\title{Batch Process Winsteps Files}
\usage{
batchWinsteps(batchName, files = NULL, dir = getwd(), outFileNames = NULL,
  pattern = "Cntrl")
}
\arguments{
\item{batchName}{The name of the batch file to be saved.}

\item{files}{Optional vector of the names of all Winsteps control files.
Note that only the name is requred, not the extension (.txt is assumed). If
argument is not supplied, the function will search the working directory
files with a common string pattern.}

\item{dir}{Directory where multiple control files are saved. Defaults to
current directory.}

\item{outFileNames}{Character vector of output file names from Winsteps
analyses. Defaults to NULL, and the function creates names based on the
names of the control files.}

\item{pattern}{Common character pattern represented in control files in the
directory. Defaults to "Cntrl", which is what is automatically generated
when r2Winsteps is called. Note that if \code{files} argument is supplied,
\code{pattern} is ignored.}
}
\description{
This function searches the working directory for Winsteps control files and
  creates a batch processing file based on all the control files in the
  directory. Alternatively, a vector of control file names can be supplied,
  from which teh batch processing file is constructed.
}

