% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r2Winsteps.r
\name{r2Winsteps}
\alias{r2Winsteps}
\title{Write data and control files for Winsteps from R.}
\usage{
r2Winsteps(itms, dems, partialCredit = FALSE, title = "r2Winsteps",
  Ifile = TRUE, Pfile = TRUE, format = "txt")
}
\arguments{
\item{itms}{Dataframe or matrix of items responses}

\item{dems}{Dataframe or matrix of person identifiers/demographic fields}

\item{title}{Title of the analysis, which will be used in the name of the
data and text files. Defaults to "r2Winsteps".}

\item{format}{Format in which item and person files should be returned.
Takes values "txt" and "XLS" to return txt and XLS files, respectively.
Defaults to "txt". Note that "txt" must be used if subsequent calls to
batch.ifile/batch.pfile are used to read thefiles back into R.}

\item{ifile}{Logical, default to TRUE. Should item files be returned?}

\item{pfile}{Logical, default to TRUE. Should person files be returned?}
}
\value{
Control and Data file for analysis with Winsteps.
}
\description{
Function for writing files from R to Winsteps for Rasch analysis.
}

