% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runWinsteps.r
\name{runWinsteps}
\alias{runWinsteps}
\title{Run a Rasch model from R using Winsteps.}
\usage{
runWinsteps(itms, dems, keep = FALSE, ...)
}
\arguments{
\item{itms}{Dataframe or matrix of item responses to be analyzed.}

\item{dems}{Dataframe or matrix of person identifiers/demographic fields.}

\item{keep}{Logical. Should the external files used to conduct the analysis
be stored in the working directory? If TRUE, Winsteps control and data
files will be retained, as well as item and person files in .txt format,
and the .bat file used to process the data.}

\item{...}{Additional arguments passed to \code{\link{r2Winsteps}}.}
}
\value{
List containing the item and person parameters from the given
  analysis.
}
\description{
This function uses the Winsteps software to analyze data, and then reads the
  data back into R without leaving  R interface or interacting with the
  Winsteps GUI.
}
\seealso{
\code{\link{r2Winsteps}} \code{\link{batchWinsteps}}
  \code{\link{batch.pfile}} \code{\link{batch.ifile}}
}

