% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/science.R
\docType{data}
\name{science}
\alias{science}
\title{Rating scale data on students' liking for science data}
\format{A data frame with 75 rows and 29 columns.
	\describe{
		\item{item1 - item25}{Integer item responses: 0, 1, 2}
		\item{Sex}{Factor variable representing whether the respondent
				   identified as male (M) or female (F).}
		\item{LastName, FirstName, MiddleName}{Names of respondents}
}}
\usage{
science
}
\description{
These data represent the example0dat dataset from the Winsteps software.
They represent students' ratings about how much the "like" particular 
science items, with possible values 0, 1, or 2, corresponding to the
responses "dislike", "neutral", or "like".
}
\keyword{datasets}
