% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAnchor.r
\name{write.anchor}
\alias{write.anchor}
\title{Write anchor files for equating scales with common items.}
\usage{
write.anchor(rawItmsDta, eqIfile, anchorFileName = "AnchorFile",
  eqIDcol = ncol(eqIfile), eqDiffcol = "Difficulty")
}
\arguments{
\item{rawItmsDta}{Raw dataframe of items to be analyzed. Must not contain
any non-item response columns. The \code{names} of the items must match
the item identifier in the item file (second argument in the function).}

\item{eqIfile}{An item file (ifile) from a previously run model, which 
includes common items with the raw data file. The raw data file items will
be estimated relative to the estimated item difficulties of these common
items.}

\item{anchorFileName}{Name of the anchor file to be written. Defaults to 
"AnchorFile".}

\item{eqIDcol}{The column name or number in the item file that includes item 
identifiers. Defaults to the final column.}

\item{eqDiffcol}{The column that item difficulties reside in the item file. 
Defaults to "Difficulty", which is the name of the column if r2Winsteps()
was run to produce the ifile.}
}
\value{
Item anchor file for equating scales with common items.
}
\description{
This function takes two primary argument: The raw dataframe of items to be 
  analyzed, \code{rawItmsDta}, and an item file with item difficulties 
  previously estimated, \code{eqIfile}. Note that the item identifiers in 
  \code{eqIfile} must match the \code{names} of the items in 
  \code{rawItmsDta}. The function will then write an anchor file with the 
  common items specified as having a fixed item difficulty, according to 
  their value in \code{eqIfile}.
}

