% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batchIfile.r
\name{batch.ifile}
\alias{batch.ifile}
\title{Batch Read Winsteps Item Files (ifile) Into R}
\usage{
batch.ifile(dir = getwd(), pattern = "Ifile")
}
\arguments{
\item{dir}{The directory where the item files are saved.}

\item{pattern}{The common character string in all item files. Defaults to "Ifile", which is what is automatically generated when r2Winsteps is called.}
}
\value{
List of Winsteps ifiles, where each element of the list represents a different ifile.
}
\description{
This function looks at a directory, identifies all the ifiles in the directory based on a common string pattern, and returns a list with all item files (i.e., each element of the list represents a different pfile).
}

