% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batchPfile.r
\name{batch.pfile}
\alias{batch.pfile}
\title{Batch Read Winsteps Person Files (pfile) Into R}
\usage{
batch.pfile(demNameL = NULL, dir = getwd(), pattern = "Pfile",
  r2WinstepsFile = TRUE)
}
\arguments{
\item{demNameL}{Optional list containing the names of the demographic variables for pfiles.}

\item{dir}{Directory where pfiles are saved. Defaults to current working directory.}

\item{r2WinstepsFile}{Logical. Was the r2Winsteps function called to produce the pfiles? Defaults to TRUE. Note that if FALSE all demographic variables will be read into a single variable.}
}
\value{
List of person files, with each element of the lit representing a sererate person file.
}
\description{
This function looks at a directory, identifies all the pfiles in the directory based on a common string pattern, and returns a list with all person files (i.e., each element of the list represents a different pfile).
}

