% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batchPfile.r
\name{batch.pfile}
\alias{batch.pfile}
\title{Batch Read Winsteps Person Files (pfile) Into R}
\usage{
batch.pfile(demNameL = NULL, dir = getwd(), files = NULL,
  pattern = "Pfile", r2WinstepsFile = TRUE, varWidths = c(1, 5, 8, 3, 8,
  9, 7, 7, 7, 7, 7, 7, 7, 7, 6, 6), pFileNames = c("Dropped", "Entry",
  "Theta", "Status", "Count", "RawScore", "SE", "Infit", "Infit_Z", "Outfit",
  "Outfit_Z", "Displacement", "PointMeasureCorr", "Weight", "ObservMatch",
  "ExpectMatch"))
}
\arguments{
\item{demNameL}{Optional list containing the names of the demographic
variables for pfiles. Must equal the number of demographic fields being
read into R.}

\item{dir}{Directory where pfiles are saved. Defaults to current working
directory.}

\item{files}{Optional character vector of the names of the person files to be
read into R. Note that only the names of the files (not the extension) are
neccessary. The function assumes .txt files.}

\item{pattern}{A common character string in all item files. Only used if
\code{files = NULL}, Defaults to "Pfile", which is what is automatically
generated when \code{link{r2Winsteps}} is called.}

\item{r2WinstepsFile}{Logical. Was the \code{link{r2Winsteps}} function
called to produce the person files? Defaults to TRUE. Note that if FALSE
all demographic variables will be read into a single variable.}

\item{varWidths}{Width of the pfile columns. This should only need to be
modified if you have an older version of Winsteps. Note that these widths
correspond to all but the final person demographic columns.}

\item{pFileNames}{The column names for the person file being read in. Note
  that these names correspond to all but the final person demographic
  columns. Additionally, the \code{length(pFileNames)} must equal
  \code{length(varWidths)}.}
}
\value{
List of person files, with each element of the lit representing a
  separate person file.
}
\description{
This function looks at a directory, identifies all the pfiles in the
 directory based on a common string pattern, and returns a list with all
 person files (i.e., each element of the list represents a different pfile).
 Alternatively, a string vector of files to read in can be supplied. Note
 that this function assumes standard .txt person files from Winsteps have
 been exported (i.e., the files have not been modified from the standard
 Winsteps output format).
}

