% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batchWinsteps.r
\name{batchWinsteps}
\alias{batchWinsteps}
\title{Batch Process Winsteps Files}
\usage{
batchWinsteps(batchName, dir = getwd(), outFileNames = NULL,
  pattern = "Cntrl")
}
\arguments{
\item{batchName}{The name of the batch file to be saved.}

\item{dir}{Directory where multiple control files are saved. Defaults to
current directory.}

\item{outFileNames}{Character vector of output file names from Winsteps
analyses. Defaults to NULL, and the function creates names based on the
names of the control files.}

\item{pattern}{Common character pattern represented in control files in the
directory. Defaults to "Cntrl", which is what is automatically generated
when r2Winsteps is called.}
}
\description{
This function looks at a directory and creates a Winsteps batch processing
  file based on all the control files in the directory.
}

