% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.r2Winsteps}
\alias{plot.r2Winsteps}
\title{Plot the test characteristic curve}
\usage{
plot.r2Winsteps(ob, type = "TIF", theta = seq(-4, 4, 0.1),
  itemSelect = NULL, colors = NULL, legend = TRUE, store = FALSE, ...)
}
\arguments{
\item{ob}{Objects of class r2Winsteps}

\item{type}{The type of plot to produce. Valid values are "TIF", "IIFs",
"TIF/IIF", "TCC", "ICCs" "ICP", and "thresholds". These correspond to,
respectively, the test information function,  item information
functions, both the test and item information functions, the test
characteristic curve, item characteristic curves, item-category
probabilities, and the Thurstonion thresholds. Note that ICP and Thresholds
are only available for polytomous models. Defaults to "TIF".}

\item{theta}{Theta values from which the information function values are
calculated from. Defaults to a sequence from -4 to 4 by 0.1. The limits of
this range are used for the x-axes.}

\item{itemSelect}{Items to be used in the plotting. Defaults to NULL, in
which case all items will be used.}

\item{colors}{Line colors for the plot. Defaults to NULL, in which case the
\code{rainbow} function is used to create colors.}

\item{store}{Optional logical argument to return data used in plotting
(not available for all plot types). For example, when \code{store = TRUE}
and \code{type = TIF}, information under the specified theta range will be
returned.}

\item{...}{Additional arguments passed to \code{plot}}
}
\description{
Plot the test characteristic curve
}

